move connect function
