+++ crates/nfs3_client/src/lib.rs
@@ -1,3 +1,4 @@
+pub(crate) mod connect;
 pub mod error;
 pub mod io;
 pub(crate) mod mount;
@@ -6,6 +7,7 @@ pub(crate) mod nfs;
 pub(crate) mod portmapper;
 pub mod rpc;
 
+pub use connect::*;
 pub use mount::*;
 pub use nfs::*;
 pub use portmapper::*;
